Make sample plot more useable
Now the sample plot script checks whether a filename was provided on the
command line, and if not, shows the figure with pyplot; if a filename is
provided, the ordinary behavior (printing figure to filename) prevails.
The printed figure now uses tex for rendering, and even a brief preamble to
make the typeface match the rest of the document.